Added authorization + local storage in Frontend

+++ TPAWebApi/TPAWebApi/Controllers/AppointmentController.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
@@ -40,7 +41,11 @@ namespace TPAWebApi.Controllers
         [HttpGet("{userId}")]
         public IActionResult GetAppointments(int userId)
         {
+            if (!userAuthorized(Request.Headers["Authorization"], userId))
+                return Unauthorized();
+
             if (userId <= 0) return BadRequest(nameof(userId));
+
             var appointments = _unitOfWork
                                    .Appointments
                                    .Find(x => x.UserId == userId)
@@ -48,10 +53,12 @@ namespace TPAWebApi.Controllers
             return Ok(appointments.ToArray());
         }
 
-        // POST api/values
         [HttpPost]
-        public IActionResult Post(AppointmentDto appointment)
+        public IActionResult Post([FromBody]AppointmentDto appointment)
         {
+            if (!userAuthorized(Request.Headers["Authorization"], appointment.UserId))
+                return Unauthorized();
+
             if (appointment == null) throw new ArgumentNullException(nameof(appointment));
             var _appointment = _mapper.Map<AppointmentDto, Appointment>(appointment);
             _unitOfWork.Appointments.Create(_appointment);
@@ -60,10 +67,12 @@ namespace TPAWebApi.Controllers
             return Ok(appointment);
         }
 
-        // PUT api/values/5
         [HttpPut("{id}")]
-        public IActionResult Put(int id, AppointmentDto appointment)
+        public IActionResult Put(int id, [FromBody]AppointmentDto appointment)
         {
+            if (!userAuthorized(Request.Headers["Authorization"], appointment.UserId))
+                return Unauthorized();
+
             if (id <= 0) BadRequest(nameof(id));
             var _appointment = _mapper.Map<AppointmentDto, Appointment>(appointment);
             _unitOfWork.Appointments.UpdateById(id, _appointment);
@@ -72,14 +81,32 @@ namespace TPAWebApi.Controllers
             return Ok(appointment);
         }
 
-        // DELETE api/values/5
         [HttpDelete("{id}")]
-        public IActionResult Delete(int id)
+        public IActionResult Delete(int id, [FromBody]int userId)
         {
+            if (!userAuthorized(Request.Headers["Authorization"], userId))
+                return Unauthorized();
+
             if (id <= 0) return BadRequest(nameof(id));
             _unitOfWork.Appointments.DeleteById(id);
             _unitOfWork.Save();
             return Ok();
+        }
+        private static bool userAuthorized(string tokenString, int id)
+        {
+            var tokenHandler = new JwtSecurityTokenHandler();
+
+            tokenString = tokenString
+                .Replace("Bearer ", "")
+                .Replace(" ", "");
+            var token = tokenHandler.ReadJwtToken(tokenString);
+            int.TryParse(token.Claims
+                    .ToArray()[0]
+                    .ToString()
+                    .Split(':')[1]
+                , out var userId);
+
+            return userId == id;
         }
     }
 }
